requirements.txt
gitpython==3.1.41
pandas==2.2.1
matplotlib==3.8.3
seaborn==0.13.2
plotly==5.19.0
pytest==8.0.2
click==8.1.7